Generalize hustle market to action market categories
